requirements.txt
# KulPik Auto-Curation Dependencies
# Python 3.8+ required

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Environment & Config
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.9

# AI/ML APIs
cohere>=5.0.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.2
